#371 feat: 커버리지 변경
