+++ backend/build.gradle
@@ -90,7 +90,8 @@ jacocoTestReport {
 							'**/global/**',
 							'**/BaseEntity*',
 							'**/ControllerAdvice*',
-							'**/AuthorizationExtractor*'
+							'**/AuthorizationExtractor*',
+							'**/applicationrunner/**',
 					])
 				})
 		)
@@ -107,7 +108,7 @@ jacocoTestCoverageVerification {
 			limit {
 				counter = "LINE"
 				value = "COVEREDRATIO"
-				minimum = 0.85
+				minimum = 0.75
 			}
 
 			excludes = [
@@ -118,7 +119,8 @@ jacocoTestCoverageVerification {
 					'*.global.*',
 					'*.BaseEntity',
 					'*.ControllerAdvice',
-					'*.AuthorizationExtractor'
+					'*.AuthorizationExtractor',
+					'*.applicationrunner.*',
 			]
 		}
 	}
