+++ backend/src/main/java/moheng/applicationrunner/warmup/AuthWarmUpRunner.java
@@ -1,4 +1,4 @@
-package moheng.applicationrunner.warmup.member;
+package moheng.applicationrunner.warmup;
 
 import moheng.auth.application.AuthService;
 import moheng.auth.dto.RenewalAccessTokenRequest;
