+++ backend/src/main/java/moheng/applicationrunner/warmup/PlannerWarmUpRunner.java
@@ -1,6 +1,6 @@
-package moheng.applicationrunner.warmup.member;
+package moheng.applicationrunner.warmup;
 
-import moheng.member.application.MemberService;
+import moheng.planner.application.PlannerService;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.annotation.Profile;
@@ -8,21 +8,19 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 @Profile({"dev", "prod"})
-@Order(5)
+@Order(9)
 @Component
-public class MemberWarmUpRunner implements ApplicationRunner {
-    private final MemberService memberService;
+public class PlannerWarmUpRunner implements ApplicationRunner {
+    private final PlannerService plannerService;
 
-    public MemberWarmUpRunner(final MemberService memberService) {
-        this.memberService = memberService;
+    public PlannerWarmUpRunner(final PlannerService plannerService) {
+        this.plannerService = plannerService;
     }
-
 
     @Override
     public void run(ApplicationArguments args) throws Exception {
         try {
-            memberService.findByEmail("email");
-            memberService.findMemberAuthorityAndProfileImg(1L);
+            plannerService.findPlannerOrderByRecent(1L);
         } catch (Exception e) {
 
         }
